Fix hidden line numbers
The issue is that the `<code>` element needs the `overflow: scroll` (since the code block has a `max-height`). So the line numbers container, which is positioned absolutely using a negative left position is hidden.

This is fixed by changing the way the `<code>` element is positioned: not with a margin and a small size, but with the padding & a larger container.