Use the new array syntax for clike and coffeescript
This patch uses the new array syntax for the `clike` language and for
`coffeescript`. For `clike` it fixes a minor bug, whereby The character
immediately in front of a single line comment is
highlighted as a comment as well.

something// something

This is because both definitions for single and multiline comments
have to be matched with a single regex and the `lookbehind` parameter
can only be applied to the first captured string.

For `coffeescript` the previously used `block-comment` token name is no
longer necessary.