Fix Java operators to correctly highlight examples.html
This patch basically adds a lot of `=?` at the end of operators that
were missing. The order of the `>=` and `<=` operator was wrong. It also
changes the order of the lookbehind pattern, so that the start of the
line is matched first. This last change fixes a lot of the bugs on the
`example.html` page.

The design of Prism makes it seem as if the regex multi line flag has no
effect, because strings are continuosly split up before the regexes are
applied. In this case it is necessary so that `^` is interpreted as the
start of a line instead of the start of the document.