Fix line-numbers for PHP in markup
The problem is, that inline php code can be anywhere inside of a HTML
element. To properly highlight this the php language implementation
temporarily replaces all inline php code in `env.code` with something
like `{{PHP0}}`. After the markup highlighting is finished it replaces
it again with the real code, but it leaves `env.code` corrupted.
Since the line-numbers plugin uses `env.code` to count the number of
lines, the result will be broken.

So to fix this, it must be ensured, that `env.code` is restored before
the line-numbers plugin uses it.