Partial solution for the "Comment-like substrings"-problem
This patch introduces a new attribute called `greedy`. The
attribute is a simple boolean flag. If there is no match
for a greedy pattern it can concatenate the next two tokens
into a single string and try to match on this string again.
If a match is found on the second attempt, then the old tokens are
deleted and replaced by the new match.

This solves the "Comment-like substrings"-problem for exactly one
comment at very little cost. With this patch the following code is
highlighted correctly:

"foo /* bar */ baz"; "foo // bar";
/lala"test"sdf/;

This approach fails if there are more than one comments inside the
string:

"foo /* bar */ baz /* bar */ baz";

Signed-off-by: Andreas Rohner <andreas.rohner@gmx.net>