Fix regression in Bash strings
In a recent patch the regex for single-quoted and double-quoted
strings of the Bash language was split into two separate regexes
to prevent the highlighting of variables inside of single-quoted
strings.

This causes a bug whenever a double-quoted string apears inside a
single-quoted string, because the double-quoted string is matched
first.

'"foo"'

The same problem exists with the newly introduced Here-Documents.
This patch fixes the problem by matching Here-Documents first and
merging the regexes for single-quoted and double-quoted
strings again.

This patch also adds testcases for this, to prevent future
regressions.