Fix built-in commands and functions for the Bash language
Keywords and functions should not start with a boundary character,
because this leads to lots of false matches. Bash actually only
allows a few boundary characters for commands and functions. This
patch replaces the \b with an explicit list of allowed characters.
This improves the highlighting of pathnames enourmously.

Some examples of invalid highlighting this patch fixes:

cat /etc/passwd
cat ./if/else/something-file.txt

This patch also adds the built-in commands ":" and ".".

: NOOP-command
. Source command