Add support for arithmetic environments for the Bash language
Currently arithmetic operators are wrongly matched all over the
place. For example the dash in `ls -la` is highlighted as
a minus operator, but for Bash it is only a string. Another
example would be a simple path like this `cat 801-file.txt`. Here
`801` is wrongly highlighted as a number and `-` is highlighted as
an operator.

Bash only supports numbers and arithmetic operators inside of a so called
arithemtic environment `$((3 + 4))`. By supporting this and
limiting numbers and arithmetic operators to arithemtic
environments, a lot of wrongly highlighted stuff can be avoided.