Parse Settings plugin
This plugin is not intended to be used on its own. It is a support
plugin for other plugins and only included as a dependency.

All ancestors of the current element are automatically scanned
for data-*-attributes and classes. These attributes and classes
are then parsed into key-value pairs and inserted into the
env.settings object. An element can inherit the data-*-attributes
from its ancestors.

Other plugins can use the collected settings from the Prism
environment (env.settings).