New Theme for PrismJS - Solarized Light
I quite like the solarized themes - they are pretty accessible and a lot of work was done on created something that is easy on the eyes. I ported Solarized Light for PrismJS. It would be pretty cool if the solarized themes were included with PrismJS! However, after some research, I couldn't quite see any attribution rules so I provided a link to the original creator of the solarized formats in a comment at the top of the css file.