Add property 'aliasTitles' to components.js
This is necessary, because a language that has a lot of aliases
like markup (HTML, SVG, XML, MathML), is only added to the
languageMap once as "Markup". So any plugin that depends on the
languageMap is limited by that.

This patch allows the definition of additional titles for
languages, depending on the alias that is used.