Plugins: Toolbar & Copy to Clipboard (#891)
* Add prism-toolbar plugin

This plugin exposes a `registerButton` method, which other
plugins can use to add buttons to the toolbar. Comes with
styles.

* Add demo file for toolbar plugin

Registers a "Hello World!" tag with the toolbar.

* Make `toolbar.registerButton` polymorphic

This allows developers to provide either a callback or an object
with a `text` string and an optional `onClick` function to create
a new button.

* Add Toolbar & Copy to Clipboard to components.js

* Add Copy to Clipboard plugin

* Switch `innerHTML` to `textContent`

This ensures additional HTML can't be passed to the toolbar
via the `text` property, ensuring a consistent display for the
buttons.

* Use `call` to bind `this` to the `onClick` method

This provides access to the clicked element, which is what `this`
is usually bound to on event listeners.

* Add hover animation to toolbar

* Add drop shadow to toolbar buttons

* Add `clipboard` to `optionalDependencies`

This will install Clipboard.js when installing from `npm`, but
won't fail the build if the installation of Clipboard.js fails.

* Load Clipboard.js from CDN if not present

* Display plugin code using data-src

* Recompile prism-toolbar

* Update Show Languages to be a Toolbar button

Show Languages now registers a callback with the toolbar
plugin to return an element with the language in it.

* Add basic HTML API & documentation

The Toolbar will now be able to read a `data-label` attribute
and add it to the code snippet.

* Switch a -> button when only providing onClick

Also adds a `url` property which creats an anchor tag and sets
the href. Adds some styles to override the button defaults.

* Add support for data-url to create anchor tag

This allows the HTML API to create links in the Toolbar.

* Update toolbar to allow order controlled via HTML

Uses a data-attribute on the `body` tag to update the order,
should the user choose to do so.

* Allow template element to provide content to label

This provides one of several options a user can implement in order to
get a custom button.

Also fixes some bugs in the documentation.

* Fix bug when combined with the autoloader plugin

The autoloader will rehighlight the element after the language arrives.
This means the complete hook can run multiple times. Without a check,
multiple toolbars can get added to an element.