Add support for IchigoJam BASIC
Hi. This PR adds support for [IchigoJam](https://ichigojam.net/), which is a board that using its own BASIC language, so I was not extending this from `basic`, but submit as a standalone language.

If there're anything necessary to change, please reply to me and I'll respond to you ASAP.