Auto-detect command lines if 'data-filter' attribute present and not empty.
This change essentially performs the same operation as the existing code which processes the 'data-output'
atrribute values, but in an automatic fashion.

When the 'data-filter' attribute is found in the <pre> element with a non-empty value, visit each line of the
<code> content. If the line starts with the value found in the 'data-prompt' attribute, treat it as a command
and wrap the text in <span class="command-line-command">. Otherwise, treat the line as command output and strip
off any data-user, data-host, or data-prompt attributes, just like the code which performs filter based on line
numbers and ranges.