Fixed Previewers bug #1496 (#1497)
The problem was a bug unique to Firefox which causes offsetTop of span.token
to always be 0 under certain conditions. It only occurred together with of the
line-number plugin because setting position: relative for the pre > code elements
triggers that FF bug.

The offset is now calculated using bounding boxes.

Tested in Chrome, Edge, IE, FF, and Opera.

Fixes #1496.