Call `insert-before` hook properly (#1709)
Fixes #1706 (also includes a demo).

The problem was that the `insert-before` hook was not called if `grammar` is undefined.
This also fixes that `callback` was not called in that case and that `env` of the `complete` hook did not have a `highlightedCode` property.

All other hooks are unchanged. The `insert-before` hook is only used by the Command line plugin, so there should be no other side effects.