Website: Added and updated documentation (#2654)
I updated a few of the entries in the "Full list of features" section (consistent formatting, a bit of grammar, nothing big). For some reason, we had both

- "Highlights embedded languages (e.g. CSS inside HTML, JavaScript inside HTML)" and 
- "Highlights nested languages (CSS in HTML, JavaScript in HTML)"

in there. They are sort of the same, so I removed one.

I also added a note regarding escaping and gave the manual highlighting section its own heading.