Initial bit of prettier, with recommended .prettierrc
Note: I only ran prettier on `src/core`. The reason for this is that this prettier config will mangle some code.

It is not the fault of a single prettier plugin, or the prettier version. Instead, it is when 3 plugins are used together that this issue happens.

It is kind of difficult to describe - you can see the error on the bash language defintion

`npx prettier src/languages/prism-bash.ts --write`

Additionally, I did not want to work super hard on forming eslint to the prettier config until we committed to this approach. I would love feedback.