Ensuring we can import from ESM and CJS modules, including typing
I duplicated the bundler declarations because I could not figure out how to get the options of declarationDir, output.dir and write({dir}) to all work together.

However it is also helpful to compile the ESM to CJS. I am not sure if this is the best way to do this.