Fix the `type-check` action (#3893)
* Update `typescript`

* Ignore the non-existent capturing group error TS(2532).

We ignore the non-existent capturing group TS error (by expecting it) where it’s irrelevant and preserve the existing `eslint-disable` rules. There was no way to stack the existing ESLint rule with the new one (they should be applied to the same line), so we placed the ESLint rule on the same line it disables.