Remove `postinstall` hook
In the postinstall hook, we call the build script. We are building Prism v1 with Gulp. Gulp is a dev dependency. The dev dependencies are never installed if a package is used as a dependency (no matter what kind of dependency). Because of that, every time we do `npm install` when working with the website repo, the installation process fails (because Gulp can’t be found).